Fix to change FAR id from uint8 to uint32_t and fixing the log to print error instead of accessing null pointer

+++ pfcpiface/messages.go
@@ -382,7 +382,7 @@ func (pc *PFCPConn) handleSessionModificationRequest(upf *upf, msg message.Messa
 			return sendError(err)
 		}
 
-		f, err := session.RemoveFAR(uint8(farID))
+		f, err := session.RemoveFAR(farID)
 		if err != nil {
 			return sendError(err)
 		}
